requirements-test.txt
pytest>=8.0.0
pytest-django>=4.8.0
